+++ src/CleanRegex/Internal/Match/Stream/IntegerStream.php
@@ -1,13 +1,12 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Match\Stream;
 
-use TRegx\CleanRegex\Exception\IntegerFormatException;
-use TRegx\CleanRegex\Exception\IntegerOverflowException;
 use TRegx\CleanRegex\Exception\InvalidIntegerTypeException;
+use TRegx\CleanRegex\Internal\Match\Stream\Number\IdentityNumber;
+use TRegx\CleanRegex\Internal\Match\Stream\Number\IntableNumber;
+use TRegx\CleanRegex\Internal\Match\Stream\Number\Number;
+use TRegx\CleanRegex\Internal\Match\Stream\Number\StreamNumber;
 use TRegx\CleanRegex\Internal\Numeral\Base;
-use TRegx\CleanRegex\Internal\Numeral\NumeralFormatException;
-use TRegx\CleanRegex\Internal\Numeral\NumeralOverflowException;
-use TRegx\CleanRegex\Internal\Numeral\StringNumeral;
 use TRegx\CleanRegex\Internal\Type\ValueType;
 use TRegx\CleanRegex\Match\Details\Intable;
 
@@ -28,32 +27,30 @@ class IntegerStream implements Upstream
 
     public function all(): array
     {
-        return \array_map([$this, 'parse'], $this->upstream->all());
+        return \array_map([$this, 'parsedNumber'], $this->upstream->all());
     }
 
     public function first(): int
     {
-        return $this->parse($this->upstream->first());
+        return $this->parsedNumber($this->upstream->first());
     }
 
-    private function parse($value): int
+    private function parsedNumber($value): int
+    {
+        return $this->number($value)->toInt();
+    }
+
+    private function number($value): Number
     {
         if (\is_int($value)) {
-            return $value;
+            return new IdentityNumber($value);
         }
         if ($value instanceof Intable) {
-            return $value->toInt();
+            return new IntableNumber($value, $this->base);
         }
-        if (!\is_string($value)) {
-            throw InvalidIntegerTypeException::forInvalidType(new ValueType($value));
+        if (\is_string($value)) {
+            return new StreamNumber($value, $this->base);
         }
-        $number = new StringNumeral($value);
-        try {
-            return $number->asInt($this->base);
-        } catch (NumeralOverflowException $exception) {
-            throw IntegerOverflowException::forStream($value, $this->base);
-        } catch (NumeralFormatException $exception) {
-            throw IntegerFormatException::forStream($value, $this->base);
-        }
+        throw InvalidIntegerTypeException::forInvalidType(new ValueType($value));
     }
 }