+++ src/CleanRegex/Internal/Match/Stream/Number/IdentityNumber.php
@@ -1,0 +1,18 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Match\Stream\Number;
+
+class IdentityNumber implements Number
+{
+    /** @var int */
+    private $integer;
+
+    public function __construct(int $integer)
+    {
+        $this->integer = $integer;
+    }
+
+    public function toInt(): int
+    {
+        return $this->integer;
+    }
+}