+++ test/Fakes/CleanRegex/Match/Details/ConstantInt.php
@@ -7,14 +7,20 @@ class ConstantInt implements Intable
 {
     /** @var int */
     private $integer;
+    /** @var int */
+    private $expectedBase;
 
-    public function __construct(int $integer)
+    public function __construct(int $integer, int $expectedBase)
     {
         $this->integer = $integer;
+        $this->expectedBase = $expectedBase;
     }
 
     public function toInt(int $base = null): int
     {
-        return $this->integer;
+        if ($this->expectedBase === $base) {
+            return $this->integer;
+        }
+        throw new \AssertionError('Failed to assert the expected base');
     }
 }